moved changes to V19; moved/fixed incorrectly placed or documented V18 changes
