+++ synthesizer/process/src/stack/deploy.rs
@@ -214,7 +214,7 @@ impl<N: Network> Stack<N> {
             };
             // Retrieve the variable limit.
             let variable_limit = verifying_key.num_variables();
-            // If the consensus version is >= V17, set the density limit, accounting for one non-zero entry (with value 1) added to
+            // If the consensus version is >= V18, set the density limit, accounting for one non-zero entry (with value 1) added to
             // each of A, B and C in order to make the Varuna zerocheck hiding.
             let non_zero_limit = if consensus_version >= ConsensusVersion::V18 {
                 let info = verifying_key.circuit_info;
