+++ synthesizer/src/vm/tests/mod.rs
@@ -38,10 +38,10 @@ mod test_v15;
 mod test_v16;
 
 #[cfg(feature = "test")]
-mod test_v17;
+mod test_v18;
 
 #[cfg(feature = "test")]
-mod test_v18;
+mod test_v19;
 
 #[cfg(feature = "test")]
 use super::*;
