+++ synthesizer/src/vm/tests/test_v19/mod.rs
@@ -13,26 +13,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Tests for block-wide synthesis limits.
-mod blockwide_synthesis_limit;
+// Tests that the translation-marked variants of Input and Output are checked correctly.
+mod translated_type_checks;
 
 use super::*;
 
-use std::{collections::HashSet, sync::Arc};
-
-use crate::vm::test_helpers::*;
-
-use console::{
-    account::{Address, PrivateKey},
-    network::ConsensusVersion,
-    prelude::FromStr,
-    program::Value,
-};
-
-use snarkvm_ledger_block::{Deployment, Solutions, Transaction};
-use snarkvm_ledger_narwhal_subdag::test_helpers::subdag_with_cert_count;
-use snarkvm_synthesizer_program::{FinalizeGlobalState, Program};
-use snarkvm_synthesizer_snark::VerifyingKey;
-use snarkvm_utilities::{TestRng, try_vm_runtime};
-
 use super::test_v14::add_and_test_with_costs;
+use crate::vm::test_helpers::*;
